+++ recurr-backend/src/plaid/mod.rs
@@ -1,10 +1,10 @@
-use recurr_core::{Account, Transaction};
 use reqwest::header::{HeaderMap, HeaderValue};
 use serde::{Deserialize, Serialize};
 
 pub mod accounts;
 pub mod institutions;
 pub mod link;
+pub mod transactions;
 
 #[derive(Debug, thiserror::Error)]
 pub enum Error {
@@ -28,7 +28,7 @@ impl serde::Serialize for Error {
 #[derive(Serialize)]
 pub struct PlaidRequest {
     endpoint: String,
-    data: serde_json::Value,
+    data: Option<serde_json::Value>,
 }
 
 #[derive(Serialize, Debug)]
@@ -49,64 +49,6 @@ pub struct PublicTokenExchangeResponse {
 }
 
 #[tauri::command]
-pub async fn get_transactions(
-    auth_key: &str,
-    access_token: &str,
-    account_ids: Vec<String>,
-) -> Result<(Vec<Account>, Vec<Transaction>), Error> {
-    let mut authorization = String::from("Bearer ");
-    authorization.push_str(auth_key);
-
-    let mut headers = HeaderMap::new();
-    headers.insert("Authorization", authorization.parse().unwrap());
-    headers.insert("Content-Type", HeaderValue::from_static("application/json"));
-
-    #[derive(Serialize)]
-    struct Options {
-        account_ids: Vec<String>,
-    }
-
-    #[derive(Serialize)]
-    struct Request {
-        access_token: String,
-        options: Option<Options>,
-
-        start_date: String,
-        end_date: String,
-    }
-
-    let data = serde_json::to_value(Request {
-        access_token: access_token.to_string(),
-        options: Some(Options { account_ids }),
-        start_date: "2022-01-01".to_string(),
-        end_date: "2023-01-01".to_string(),
-    })?;
-
-    let req = PlaidRequest {
-        endpoint: "/transactions/get".to_string(),
-        data,
-    };
-
-    let client = reqwest::Client::new();
-    let res = client
-        .post(env!("PLAID_URL"))
-        .json(&req)
-        .headers(headers)
-        .send()
-        .await?
-        .error_for_status()?;
-
-    #[derive(Debug, Deserialize)]
-    struct Response {
-        accounts: Vec<Account>,
-        transactions: Vec<Transaction>,
-    }
-
-    let json: Response = res.json().await?;
-    Ok((json.accounts, json.transactions))
-}
-
-#[tauri::command]
 pub async fn item_public_token_exchange(
     auth_key: &str,
     public_token: &str,
@@ -124,7 +66,7 @@ pub async fn item_public_token_exchange(
 
     let req = PlaidRequest {
         endpoint: "/item/public_token/exchange".to_string(),
-        data,
+        data: Some(data),
     };
 
     let client = reqwest::Client::new();
